Add recursive controllers and refactoring
