+++ controllers/API/announcement-item.js
@@ -54,5 +54,4 @@ router.get('/announcement/item', function(req, res) {
 
 });
 
-
 module.exports = router;
