+++ controllers/auth/logout.js
