+++ controllers/main/overview-pages/lecturerpage.js
@@ -1,6 +1,6 @@
 var express = require('express');
 var router = express.Router();
-var data = require('../config/database.js');
+var data = require.main.require('./config/database.js');
 	
 router.get('/lecturerpage', data.isLoggedIn,function(req,res){
       data.db.lecturers.find(function(err,docs){
